Replace BIGSERIAL with SERIAL

+++ sql/1-createdb.sql
@@ -7,7 +7,7 @@ DROP TABLE IF EXISTS message CASCADE;
 DROP TYPE IF EXISTS SCHEDULE_T CASCADE;
 
 CREATE TABLE employer (
-  employer_id BIGSERIAL PRIMARY KEY,
+  employer_id SERIAL PRIMARY KEY,
   title VARCHAR(1000) NOT NULL
 );
 
@@ -19,8 +19,8 @@ CREATE TYPE SCHEDULE_T AS ENUM (
 );
 
 CREATE TABLE vacancy (
-  vacancy_id BIGSERIAL PRIMARY KEY,
-  employer_id BIGINT REFERENCES employer(employer_id) NOT NULL,
+  vacancy_id SERIAL PRIMARY KEY,
+  employer_id INTEGER REFERENCES employer(employer_id) NOT NULL,
   title VARCHAR(1000) NOT NULL,
   city VARCHAR(100) NOT NULL,
   salary INT4RANGE,
@@ -30,13 +30,13 @@ CREATE TABLE vacancy (
 );
 
 CREATE TABLE applicant (
-  applicant_id BIGSERIAL PRIMARY KEY,
+  applicant_id SERIAL PRIMARY KEY,
   name VARCHAR(1000) NOT NULL
 );
 
 CREATE TABLE resume (
-  resume_id BIGSERIAL PRIMARY KEY,
-  applicant_id BIGINT REFERENCES applicant(applicant_id) NOT NULL,
+  resume_id SERIAL PRIMARY KEY,
+  applicant_id INTEGER REFERENCES applicant(applicant_id) NOT NULL,
   title VARCHAR(1000) NOT NULL,
   city VARCHAR(100) NOT NULL,
   salary INT4RANGE,
@@ -46,14 +46,14 @@ CREATE TABLE resume (
 );
 
 CREATE TABLE application (
-  application_id BIGSERIAL PRIMARY KEY,
-  resume_id BIGINT REFERENCES resume(resume_id) NOT NULL,
-  vacancy_id BIGINT REFERENCES vacancy(vacancy_id) NOT NULL
+  application_id SERIAL PRIMARY KEY,
+  resume_id INTEGER REFERENCES resume(resume_id) NOT NULL,
+  vacancy_id INTEGER REFERENCES vacancy(vacancy_id) NOT NULL
   );
 
 CREATE TABLE message (
-  message_id BIGSERIAL PRIMARY KEY,
-  application_id BIGINT REFERENCES application(application_id) NOT NULL,
+  message_id SERIAL PRIMARY KEY,
+  application_id INTEGER REFERENCES application(application_id) NOT NULL,
   created TIMESTAMP NOT NULL,
   applicant_to_employer BOOLEAN NOT NULL,
   text TEXT
